refactor table component
