+++ ui/app/arcade/page.tsx
@@ -1,4 +1,4 @@
-import Background from "@/components/Background";
+import Background from "@/components/background";
 
 export default function Arcade() {
   return (
