+++ ui/components/game-console/arcade/display-value.tsx
