+++ ui/components/game-console/arcade/game-console.tsx
